Prevented clients receiving UDP broadcast datagrams from themselves
This was done because we don't want UDP datagrams from ourselves. Why?
Because it causes weirdness, such as having ourselves as an available
connected client to select from.

While having that "bug" was a good feature for testing, it's bad for
normal operation because it would cause weird problems in the future.