Rewrote some of the TCP receiving code
TCP receiving code was bad. What it did was set the .WAV stream equal to
the TCP stream, which is not how you should be doing this.

At the beginning of the receive of a TCP connection, client will accept
and flush the .WAV stream. It will then read bytes in from the TCP
stream and write them to the WAV stream.