Big File Freeze Fix
Attempting to fix the cause of the program freezing when trying to send
a large ( F > ~1024 KB) file.

What was changed:
The file reading was reading only 1 byte per loop iteration and calling
a bunch of functions and doing a bunch of comparisons during each loop
iteration.
My guess is that this was ridiculously inefficient and causing the
program to lock up on my machines.
To fix this possible cause, I changed it so that during file read, the
program would read 1024 (1 KB) bytes each loop iteration.