[FINAL PRODUCT] Fixed progress bar not updating.
Was due to a division error: Integer division will always round down,
and thus the division result was 0. 0*100 was always 0, until the file
position was equal to the file length, in which case the result was
1*100.

This explains why the file send progress bar would not update until the
end, and would update VERY fast.

This is now fixed, by adding necessary type casts.